Updated Docker files for Ubuntu 20.04
Replaced the previous Docker setup files that were supporting the run of UPSET in a CentOS 7 environment, so that the user can run UPSET in Ubuntu 20.04. More information regarding the setup of the Docker environment will be added to the UPSET documentation in a future commit

WARNING: Current Docker setup supports only x86_64 architectures!

The release of the next version of UPSET will include the executable for UPSET 20.04 environments along with other features.