README and Documentation Update
The UPSET documentation is updated to include instructions regarding running UPSET inside a Docker environment as well as information regarding the supported SET mitigation techniques (and the corresponding TCL commands).

Moreover, the repository README is updated to include more information related to the UPSET project, as well as a DEMO and Related Publications section.