Solved issues regarding Docker Setup
Resolved some issues with the Docker setup. Specifically, the problem was that the relative path of the INSTALL directory was specified in the Makefile. However, Docker requires absolute paths, which caused the UPSET not to be found. Additionally, in the updated version, the overall data folder is provided in the Docker container, allowing the user to save reports, etc., and to open or edit them on the native machine.

Also, updated the instructions regarding the Docker setup, both in the README file and the documentation as well.

Co-Authored-By: D.Vali <56717600+dvaliantzas@users.noreply.github.com>
Co-Authored-By: zazatisn <190748969+zazatisn@users.noreply.github.com>